+++ pos-inventory/src/main/java/com/positivity/inventory/internal/dto/InventoryLedgerEntryResponse.java
@@ -49,11 +49,11 @@ public class InventoryLedgerEntryResponse {
             example = "12",
             requiredMode = REQUIRED)
     @NotNull
-    Integer changeInQuantity;
+    BigDecimal changeInQuantity;
 
     @Schema(description = "Running quantity after this entry was applied", example = "120", requiredMode = REQUIRED)
     @NotNull
-    Integer quantityAfter;
+    BigDecimal quantityAfter;
 
     @Schema(
             description = "Unit cost associated with this movement, if recorded",
