+++ pos-inventory/src/main/java/com/positivity/inventory/internal/exception/ReturnQuantityExceededException.java
@@ -1,10 +1,11 @@
 package com.positivity.inventory.internal.exception;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public class ReturnQuantityExceededException extends RuntimeException {
 
-    public ReturnQuantityExceededException(UUID skuId, int requested, int available) {
+    public ReturnQuantityExceededException(UUID skuId, BigDecimal requested, BigDecimal available) {
         super("Return quantity exceeds available consumed quantity for skuId "
                 + skuId
                 + ": requested="
